chore: add new interface file for lens components
